backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pypdf
python-dotenv